Isolate tests which touch the Warehouse module.
Ideally, these global configuration variables would be abstracted so
that they could be injected into the classes that need them, instead
of being tied to the Warehouse module in this way. Functionally,
though, the tests would still need this reset capability, since
DataMapper's model list would still be a global variable (as is
reasonable enough for most uses of DataMapper).