A simple interface for updating .pgpass files.
PostgreSQL's command line utilities require either that the password be provided interactively
or that it be stored in a ".pgpass" file with a certain format and permissions. Since we have
the password in bcdatabase already, requiring the user to create and update this file will be
a source of error.