Ensure that all MDES variables refs become usable belongs_tos.
Some reference types in the MDES do not follow the _id suffix convention for ID references.
This change ensures that their models are generated with a suffixed association name so that
there is no collision between the FK column and the association name (both of which become
methods on model instances).