Restore foreign key creation. Closes #1639.
No amount of hackery would encourage dm-migrations and dm-constraints to
generate the FKs in anything other than the :default repository. This meant
that I had to change the behavior of
`DatabaseInitializer#set_up_repository(:both)` so that the `:default` repo
is the working repo when both are connected. This is fine for the warehouse's
internal purposes, and external consumers aren't intended to use `:both`.