Work around CI spec failure with RSpec 2.10.
CI Reporter wraps the IO streams, capturing input and delegating most other
methods to the wrapped stream. Using RSpec 2.10 and CI Reporter 1.6.6, the
"actual" in this check is the wrapped stream while the "expected" is the
wrapper, causing it to fail. It does not fail with RSpec 2.9 and the same CI
Reporter, but I cannot find a change in RSpec 2.10 that would explain the
different behavior.

In any case, this workaround is equivalent for testing purposes.