Structure for tracking FK refs in-memory. #2013.
This first pass is fairly simple. It keeps all encountered primary keys in
memory for the duration of the process. Its only nod towards efficiency is
that, for any unsatisfied foreign keys it encounters, it only holds onto
enough information to re-check them at the end of the process instead of the
entire records.