Remove autoload for MDES 2.0 models.
The fact that there was one was inconsistent with the other MDES versions.
Configuration can manage loading the models for the designated MDES version
and should ensure that model loading is done in a way that does not confuse
DataMapper. This should also make it harder to accidentally have multiple
versions of the MDES models loaded in the same process accidentally.