Introduce separate `Filters` module. #3940.
The motivation is to re-use the functionality in
`Filters::CompositeFilter` (nee `Transformers::Filters`) between
`EnumTransformer` and `Contents`. Since we will eventually be adding
filters which could work in either context, moving the existing
filters to a new namespace makes sense also.

This commit also adds the new `Filters` module to the search path for
unknown constants when reading configuration files. This should ensure
that moving the filters has no impact on deployed configurations,
since they all use the unqualified name.