Filters for changing prefixes on PKs. Closes #3968.
These filters add/remove a prefix to/from all records of a given type.
The intent is that one would be used on import into the warehouse
and the other would be used on XML generation. This will allow the
system to contain prefixed IDs while preserving the original IDs for
VDR submission.