Take transformer runs out of transactions.  #4315.
We are currently running a (mad?) experiment to see if we can relax MDES
constraints while still submitting data to the VDR.  (It's understood
that the generated data will not pass schema validations in the
Workbench.)

To make this work, we need to relax all DM validations *and* disable the
per-transformer transaction, as it's entirely possible that some errors
that would ordinarily be caught by validations will now interact badly
with other non-(non-NULL) constraints.  (Column length checks, for
example.)  By removing the per-transformer transaction, we ensure that
i.e. invalid INSERTs are not ignored until the end of the transaction,
because that outer transaction has been removed.